Split timeline.dart into multiple files
